+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/content/cogwheel_chain/graph/PathedCogwheelNode.java
@@ -5,15 +5,17 @@ import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.phys.Vec3;
 
-public record PathedCogwheelNode(int side, boolean isLarge, Direction.Axis rotationAxis, BlockPos localPos) {
+public record PathedCogwheelNode(int side, boolean isLarge, Direction.Axis rotationAxis, BlockPos localPos,
+                                 boolean offsetForSmallCogwheel) {
 
     public PathedCogwheelNode(final PlacingCogwheelNode partialNode, final int side) {
-        this(side, partialNode.isLarge(), partialNode.rotationAxis(), partialNode.pos());
+        this(side, partialNode.isLarge(), partialNode.rotationAxis(), partialNode.pos(), partialNode.hasOffsetForSmallCogwheel());
     }
 
     public void write(final CompoundTag nodeTag) {
         nodeTag.putBoolean("Side", side == 1);
         nodeTag.putBoolean("IsLarge", isLarge);
+        nodeTag.putBoolean("OffsetForSmallCogwheel", offsetForSmallCogwheel);
         nodeTag.putInt("OffsetX", localPos.getX());
         nodeTag.putInt("OffsetY", localPos.getY());
         nodeTag.putInt("OffsetZ", localPos.getZ());
@@ -29,7 +31,16 @@ public record PathedCogwheelNode(int side, boolean isLarge, Direction.Axis rotat
                 nodeTag.getInt("OffsetZ")
         );
         final Direction.Axis rotationAxis = Direction.Axis.values()[nodeTag.getInt("RotationAxis")];
-        return new PathedCogwheelNode(side, isLarge, rotationAxis, offset);
+
+
+        final boolean offsetForSmallCogwheel;
+        if (nodeTag.contains("OffsetForSmallCogwheel")) {
+            offsetForSmallCogwheel = nodeTag.getBoolean("OffsetForSmallCogwheel");
+        } else {
+            offsetForSmallCogwheel = !isLarge;
+        }
+
+        return new PathedCogwheelNode(side, isLarge, rotationAxis, offset, offsetForSmallCogwheel);
     }
 
     public float sideFactor() {